+++ code/algorithms.cpp
@@ -109,11 +109,11 @@ void run_model(hplus::instance& inst, hplus::environment& env, hplus::statistics
     stopchk();
 
     if (env.alg == HPLUS_CLI_ALG_IMAI)
-        imai::build_cpx_model(cpxenv, cpxlp, inst, env, log);
+        imai::build_cpx_model(cpxenv, cpxlp, inst, env, log, stats);
     else if (env.alg == HPLUS_CLI_ALG_RANKOOH)
-        rankooh::build_cpx_model(cpxenv, cpxlp, inst, env, log);
+        rankooh::build_cpx_model(cpxenv, cpxlp, inst, env, log, stats);
     else if (env.alg == HPLUS_CLI_ALG_DYNAMIC_TIME)
-        rankooh_dynamic::build_cpx_model(cpxenv, cpxlp, inst, env, log);
+        rankooh_dynamic::build_cpx_model(cpxenv, cpxlp, inst, env, log, stats);
     stopchk();
 
     // time limit
@@ -146,7 +146,10 @@ void run_model(hplus::instance& inst, hplus::environment& env, hplus::statistics
 
     CPX_HANDLE_CALL(log, CPXmipopt(cpxenv, cpxlp));
 
+    stats.nnodes = CPXgetnodecnt(cpxenv, cpxlp);
+
     if (parse_cpx_status(cpxenv, cpxlp, env, log)) {  // If CPLEX has found a solution
+        CPX_HANDLE_CALL(log, CPXgetbestobjval(cpxenv, cpxlp, &stats.lb));
         if (env.alg == HPLUS_CLI_ALG_IMAI)
             imai::store_cpx_sol(cpxenv, cpxlp, inst, log);
         else if (env.alg == HPLUS_CLI_ALG_RANKOOH)