+++ code/exact/rankooh_dynamic.cpp
@@ -2,7 +2,7 @@
 
 void rankooh_dynamic::build_cpx_model([[maybe_unused]] CPXENVptr& cpxenv, [[maybe_unused]] CPXLPptr& cpxlp,
                                       [[maybe_unused]] const hplus::instance& inst, [[maybe_unused]] const hplus::environment& env,
-                                      [[maybe_unused]] const logger& log) {
+                                      [[maybe_unused]] const logger& log, [[maybe_unused]] hplus::statistics& stats) {
     PRINT_VERBOSE(log, "Building Rankooh's dynamic model.");
 
     todo(log, "rankooh_dynamic::build_cpx_model");  // TODO