+++ Changelog.md
@@ -25,6 +25,24 @@ HOW I DO VERSIONING:
 <!-- ### :curly_loop: Other -->
 
 
+## [1.3.0] - 2025/03/27
+### Tested all features added in 1.2.*
+<!-- ### :warning: Known issues -->
+<!-- ### :ballot_box_with_check: Fixed -->
+<!-- ### :heavy_exclamation_mark: Changed -->
+### :heavy_plus_sign: Added
+- Added status, costs and cplex execution info to statistics
+    - [~/code/exact/*](code/exact/)
+    - [~/code/algorithms.cpp](code/algorithms.cpp)
+    - [~/code/hplus_instance.hpp](code/hplus_instance.hpp)
+    - [~/code/hplus_instance.cpp](code/hplus_instance.cpp)
+    - [~/code/main.cpp](code/main.cpp)
+- Better results reading using the new statistics output
+    - [~/code/test/code/results_jobs.py](code/test/code/results_jobs.py)
+<!-- ### :x: Removed -->
+<!-- ### :curly_loop: Other -->
+
+
 ## [1.2.6] - 2025/03/27
 <!-- ### :warning: Known issues -->
 ### :ballot_box_with_check: Fixed